refactor: refactor utils code for reusing
